fix: Explicitly cast to double in Math.Round to resolve ambiguity
FanCurve.cs line 80 - Math.Round(interpolated) was ambiguous between
decimal and double overloads. Cast to double explicitly.

Fixes CS0121 error with .NET SDK 10.0.400

+++ TPFan.Shared/Models/FanCurve.cs
@@ -77,6 +77,6 @@ public record FanCurve
         var tempDiff = temperatureCelsius - lowerPoint.TemperatureCelsius;
 
         var interpolated = lowerPoint.SpeedPercent + (speedRange * tempDiff) / tempRange;
-        return (int)Math.Round(interpolated);
+        return (int)Math.Round((double)interpolated);
     }
 }
